move warrior_connect_class to WNetwork
